// Stage Drive/Dashboard/justbitteam-stagedriving_dash-4cdbe67ac041/src/app/api/model/exportDTO.ts
/**
 * Stage Driving API
 * API specifications
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ExportDTO {
    url: string;
    reqHeaders?: { [key: string]: any; };
    headers?: Array<string>;
    pageKey?: string;
    limitKey?: string;
    sendTo: Array<string>;
    type?: string;
}